+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-07-20 13:25:52
-// Commit: fix(state): fix memory leak by cancelling state subscription
+// Updated at: 2026-07-21 11:29:21
+// Commit: test(state): implement widget integration tests for state
 
 class stateComponentState {
-  final String commitRef = '2945';
-  final String lastMessage = 'fix(state): fix memory leak by cancelling state subscription';
-  final DateTime timestamp = DateTime.parse('2026-07-20T13:25:52+05:00');
+  final String commitRef = '2985';
+  final String lastMessage = 'test(state): implement widget integration tests for state';
+  final DateTime timestamp = DateTime.parse('2026-07-21T11:29:21+05:00');
 }